+++ gems/aws-sdk-qconnect/lib/aws-sdk-qconnect/endpoints.rb
@@ -418,6 +418,20 @@ module Aws::QConnect
       end
     end
 
+    class PutFeedback
+      def self.build(context)
+        unless context.config.regional_endpoint
+          endpoint = context.config.endpoint.to_s
+        end
+        Aws::QConnect::EndpointParameters.new(
+          region: context.config.region,
+          use_dual_stack: context.config.use_dualstack_endpoint,
+          use_fips: context.config.use_fips_endpoint,
+          endpoint: endpoint,
+        )
+      end
+    end
+
     class QueryAssistant
       def self.build(context)
         unless context.config.regional_endpoint