+++ gems/aws-sdk-qconnect/CHANGELOG.md
@@ -1,6 +1,11 @@
 Unreleased Changes
 ------------------
 
+1.1.0 (2023-12-01)
+------------------
+
+* Feature - This release adds the PutFeedback API and allows providing feedback against the specified assistant for the specified target.
+
 1.0.0 (2023-11-28)
 ------------------